+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- `stroke-miterlimit` will be parsed as `Number` and not as a `Length` now.
 
 ## [0.15.0] - 2018-12-13
 ### Added